Nexus-Web/requirements.txt
Flask
werkzeug
python-docx
pypdf
google-generativeai
python-pptx